+++ pkg/action/stats_test.go
@@ -1,4 +1,4 @@
-package stats
+package action
 
 import (
 	"testing"
